backend/requirements.txt
django==5.1.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
celery==5.2.7
redis==4.3.4
psycopg2-binary==2.9.5
django-cors-headers==3.14.0
pandas==2.0.3
django-environ==0.8.0

